+++ evalml/automl/automl_search.py
@@ -17,6 +17,7 @@ from .pipeline_search_plots import PipelineSearchPlots
 
 from evalml.automl.automl_algorithm import IterativeAlgorithm
 from evalml.automl.data_splitters import TrainingValidationSplit
+from evalml.automl.utils import get_default_primary_search_objective
 from evalml.data_checks import (
     AutoMLDataChecks,
     DataChecks,
@@ -30,21 +31,9 @@ from evalml.exceptions import (
     PipelineScoreError
 )
 from evalml.objectives import (
-    CostBenefitMatrix,
-    FraudCost,
-    LeadScoring,
-    MeanSquaredLogError,
-    Recall,
-    RecallMacro,
-    RecallMicro,
-    RecallWeighted,
-    RootMeanSquaredLogError,
-    get_objective,
-    get_objectives
-)
-from evalml.objectives.utils import (
-    _all_objectives_dict,
-    _print_objectives_in_table
+    get_core_objectives,
+    get_non_core_objectives,
+    get_objective
 )
 from evalml.pipelines import (
     BinaryClassificationPipeline,
@@ -57,7 +46,6 @@ from evalml.pipelines.utils import make_pipeline
 from evalml.problem_types import ProblemTypes, handle_problem_types
 from evalml.tuners import SKOptTuner
 from evalml.utils import convert_to_seconds, get_random_state
-from evalml.utils.gen_utils import classproperty
 from evalml.utils.logger import (
     get_logger,
     log_subtitle,
@@ -73,13 +61,10 @@ class AutoMLSearch:
     """Automated Pipeline search."""
     _MAX_NAME_LEN = 40
     _LARGE_DATA_ROW_THRESHOLD = int(1e5)
+    _LARGE_DATA_PERCENT_VALIDATION = 0.75
 
     # Necessary for "Plotting" documentation, since Sphinx does not work well with instance attributes.
     plot = PipelineSearchPlots
-
-    _DEFAULT_OBJECTIVES = {'binary': 'Log Loss Binary',
-                           'multiclass': 'Log Loss Multiclass',
-                           'regression': 'R2'}
 
     def __init__(self,
                  problem_type=None,
@@ -174,7 +159,7 @@ class AutoMLSearch:
         self.verbose = verbose
         self.optimize_thresholds = optimize_thresholds
         if objective == 'auto':
-            objective = self._DEFAULT_OBJECTIVES[self.problem_type.value]
+            objective = get_default_primary_search_objective(self.problem_type.value)
         objective = get_objective(objective, return_instance=False)
         self.objective = self._validate_objective(objective)
         if self.data_split is not None and not issubclass(self.data_split.__class__, BaseCrossValidator):
@@ -182,7 +167,7 @@ class AutoMLSearch:
         if self.problem_type != self.objective.problem_type:
             raise ValueError("Given objective {} is not compatible with a {} problem.".format(self.objective.name, self.problem_type.value))
         if additional_objectives is None:
-            additional_objectives = [obj for obj in get_objectives(self.problem_type) if obj not in self._objectives_not_allowed_in_automl]
+            additional_objectives = get_core_objectives(self.problem_type)
             # if our main objective is part of default set of objectives for problem_type, remove it
             existing_main_objective = next((obj for obj in additional_objectives if obj.name == self.objective.name), None)
             if existing_main_objective is not None:
@@ -239,7 +224,7 @@ class AutoMLSearch:
         self._baseline_cv_score = None
 
         if _max_batches is not None and _max_batches <= 0:
-            raise ValueError("Parameter max batches must be None or non-negative. Received {max_batches}.")
+            raise ValueError(f"Parameter max batches must be None or non-negative. Received {_max_batches}.")
         self._max_batches = _max_batches
         # This is the default value for IterativeAlgorithm - setting this explicitly makes sure that
         # the behavior of max_batches does not break if IterativeAlgorithm is changed.
@@ -247,21 +232,12 @@ class AutoMLSearch:
 
         self._validate_problem_type()
 
-    @classproperty
-    def _objectives_not_allowed_in_automl(self):
-        return {CostBenefitMatrix, FraudCost, LeadScoring,
-                MeanSquaredLogError, Recall, RecallMacro, RecallMicro, RecallWeighted, RootMeanSquaredLogError}
-
-    @classmethod
-    def print_objective_names_allowed_in_automl(cls):
-        names = [name for name, value in _all_objectives_dict().items() if value not in cls._objectives_not_allowed_in_automl]
-        _print_objectives_in_table(names)
-
     def _validate_objective(self, objective):
+        non_core_objectives = get_non_core_objectives()
         if isinstance(objective, type):
-            if objective in self._objectives_not_allowed_in_automl:
-                raise ValueError(f"{objective.name} is not allowed in AutoML! "
-                                 "Use evalml.automl.AutoMLSearch.print_objective_names_allowed_in_automl() "
+            if objective in non_core_objectives:
+                raise ValueError(f"{objective.name.lower()} is not allowed in AutoML! "
+                                 "Use evalml.objectives.utils.get_core_objective_names()"
                                  "to get all objective names allowed in automl.")
             return objective()
         return objective
@@ -404,7 +380,7 @@ class AutoMLSearch:
             default_data_split = StratifiedKFold(n_splits=3, random_state=self.random_state)
 
         if X.shape[0] > self._LARGE_DATA_ROW_THRESHOLD:
-            default_data_split = TrainingValidationSplit(test_size=0.25)
+            default_data_split = TrainingValidationSplit(test_size=self._LARGE_DATA_PERCENT_VALIDATION)
 
         self.data_split = self.data_split or default_data_split