+++ evalml/objectives/__init__.py
@@ -42,4 +42,5 @@ from .standard_metrics import (
     RecallMicro,
     RecallWeighted
 )
-from .utils import get_objective, get_objectives, print_all_objective_names
+from .utils import get_objective, get_core_objectives, get_all_objective_names, get_non_core_objectives,\
+    get_core_objective_names