+++ evalml/pipelines/components/ensemble/__init__.py
@@ -1,0 +1,4 @@
+# flake8:noqa
+from .stacked_ensemble_base import StackedEnsembleBase
+from .stacked_ensemble_classifier import StackedEnsembleClassifier
+from .stacked_ensemble_regressor import StackedEnsembleRegressor