+++ evalml/tests/model_understanding_tests/prediction_explanations_tests/test_explainers.py
@@ -159,7 +159,7 @@ def test_explain_prediction(mock_normalize_shap_values,
     pipeline.classes_ = ["class_0", "class_1", "class_2"]
 
     # By the time we call transform, we are looking at only one row of the input data.
-    pipeline._transform.return_value = pd.DataFrame({"a": [10], "b": [20], "c": [30], "d": [40]})
+    pipeline.compute_estimator_features.return_value = pd.DataFrame({"a": [10], "b": [20], "c": [30], "d": [40]})
     features = pd.DataFrame({"a": [1], "b": [2]})
     table = explain_prediction(pipeline, features, output_format=output_format, top_k=2)